requirements.txt
python-telegram-bot
pygithub
python-dotenv
requests